src/az_func/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
feedparser
azure-storage-blob
azure-storage-queue
requests
BeautifulSoup4
azure-search-documents
azure-core
python-dateutil